Issue #3 Refact: Double values in currency fields

+++ database/migrations/2023_05_23_004330_create_items_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->increments('id')->integer('id', 10);
             $table->string('name');
             $table->integer('quantity_required');
-            $table->integer('unitary_value');
-            $table->integer('total_value');
+            $table->double('unitary_value');
+            $table->double('total_value');
             $table->timestamps();
         });
     }
